+++ CHANGELOG.md
@@ -61,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Way to copy function files ([#118](https://github.com/Substra/substrafl/pull/118))
 - `download_train_task_models_by_rank` uses new function `list_task_output_assets` instead of using `value` that has been removed ([#129](https://github.com/Substra/substrafl/pull/129))
 - Python dependencies are resolved using pip compile during function registration ([#123])(https://github.com/Substra/substrafl/pull/123)).
-- BREAKING: local_dependencies are now limited to local modules or Python wheels (no support for bdist, sdist...)([#123])(https://github.com/Substra/substrafl/pull/123)).
+- BREAKING: local_dependencies is renamed local_installable_dependencies([#123])(https://github.com/Substra/substrafl/pull/123)).
+- BREAKING: local_installable_dependencies are now limited to local modules or Python wheels (no support for bdist, sdist...)([#123])(https://github.com/Substra/substrafl/pull/123)).
 
 ### Fixed
 - New dependencies copy method in Docker mode.([#130](https://github.com/Substra/substrafl/pull/130))